requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
confluent-kafka[avro]==2.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==7.2.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.16.0
numpy==1.26.4
packaging==23.2
pandas==2.3.3
pillow==10.4.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.25
streamlit==1.31.1
tenacity==8.5.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
zipp==3.23.0